Moved shared row helpers to server/lib
ref BER-3785

case-keys and the DbDate codec were defined inside gift-links but are generic
row<->domain helpers, not gift-links concerns. Moving them to server/lib lets the
upcoming members-custom-fields service reuse them without depending on gift-links.
Pure move plus import updates; no behaviour change.
